fix: remove main/Icon/stories/Styles.tsx
