+++ main/IconButton/index.tsx
@@ -5,7 +5,7 @@ import type {StyleProp, TouchableOpacityProps, ViewStyle} from 'react-native';
 
 import type {DoobooTheme} from '@dooboo-ui/theme';
 import {Icon} from '../Icon/';
-import type {IconName} from '../Icon/stories/Styles';
+import type {IconName} from '../Icon';
 import {LoadingIndicator} from '../LoadingIndicator';
 import {css} from '@emotion/native';
 import {getTheme} from '../utils';
